refactor: move StringStorageTrait to FuelPHP infrastructure
The trait depends on GuzzleHttp\Psr7\Utils (a PSR-7 implementation, not
the interface package) and was the last piece in backend/Storage that
required a framework-provided dependency to resolve. Belongs next to the
adapters that use it; backend is now Guzzle-free even transitively.
