+++ fuelphp/fuel/packages/infrastructure/classes/Storage/LocalStorage.php
@@ -11,7 +11,6 @@ use Psr\Http\Message\StreamInterface;
 use SharedKernel\Domain\Logger\LoggerInterface;
 use SharedKernel\Domain\Storage\StorageException;
 use SharedKernel\Domain\Storage\StorageInterface;
-use SharedKernel\Infrastructure\Storage\StringStorageTrait;
 
 final class LocalStorage implements StorageInterface
 {
